+++ components/scaletor/Template/index.js
@@ -1,11 +1,15 @@
 import React from 'react';
+import classNames from 'classnames';
 
 import { allTemplates } from '../../../structures/scaletor/templates';
 
-const Template = ({ id }) => {
+const Template = ({ id, templateMode }) => {
   const baseUrl = '/images/scaletor/';
   return (
-    <div className="template">
+    <div className={classNames(
+      "template",
+      templateMode && 'template__template-right',
+    )}>
       <img
         className="template__image"
         src={`${baseUrl}${allTemplates[id].imgSrc}`}